#1103 いいね機能（とさか）

+++ app/User.php
@@ -109,6 +109,11 @@ class User extends Authenticatable
     // ユーザーが特定の投稿をお気に入りに追加
     public function favorite($postId)
     {
+        $post = Post::findOrFail($postId);
+        // 投稿が自分のものであるかを確認
+        if ($post->user_id == $this->id) {
+            return false; // 自分の投稿の場合、falseを返す
+        }
         $exist = $this->isFavorite($postId); // 投稿が既にお気に入りに追加されているかどうかを確認
         if ($exist) {
             return false;
@@ -120,6 +125,11 @@ class User extends Authenticatable
     // ユーザーが特定の投稿をお気に入りから削除
     public function unfavorite($postId)
     {
+        $post = Post::findOrFail($postId);
+        // 投稿が自分のものであるかを確認
+        if ($post->user_id == $this->id) {
+            return false; // 自分の投稿の場合、falseを返す
+        }
         $exist = $this->isFavorite($postId); //投稿が既にお気に入りに追加されているかどうかを確認
         if ($exist) {
             $this->favorites()->detach($postId); //既にお気に入りに追加されている場合は、お気に入りから削除
